% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_test.R
\name{prop_test}
\alias{prop_test}
\title{prop_test}
\usage{
prop_test(db, col)
}
\arguments{
\item{db}{a data.table containing a trained column and a column of interest}

\item{col}{the column which you want to check successes for}
}
\value{
a test summary of prop.test
}
\description{
prop_test returns the test summary for a column in a given data.table
}
\examples{
test <- data.frame(trained = sample(c(0,1),100, replace = TRUE), outcome = sample(c(0,1), 100, replace = TRUE))
prop_test(test, "outcome")
}
