% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_processCols.R
\name{removeProcessCols}
\alias{removeProcessCols}
\title{removeProcessCols}
\usage{
removeProcessCols(db)
}
\arguments{
\item{db}{a table that has intermediate columns generated either by surveycto
or for surveycto}
}
\value{
a table with the process columns removed
}
\description{
removes columns that are intermediate columns not required for analysis.
}
\examples{
db <- data.frame(name_check = c("sehj","k"), subscriberid = c(1,42), sim = c("bite","shark"), sum = c(1,5))
removeProcessCols(db)
}
